Adding unit tests and some constexpr support for some string conversion routines

+++ vlr-util.test/vlr-util.test.cpp
@@ -8,7 +8,7 @@
 int main(int argc, char** argv)
 {
 	// Some tests use random numbers (using std::rand); this ensures that they are not duplicative
-	std::srand(time(0));
+	std::srand(static_cast<unsigned int>(time(0)));
 
 	testing::InitGoogleTest(&argc, argv);
 	auto nResult_GoogleTest = RUN_ALL_TESTS();
